requirements.txt
asgiref==3.5.2
Django==4.0.5
django-ckeditor==6.5.0
django-js-asset==2.0.0
Pillow==9.2.0
sqlparse==0.4.2
tzdata==2022.1
